Fix a possible edge case with KV1 capitalization handling
Basically, KV1 will cache whichever capitalization gets used first for a string, *globally*.

This previously caused issues with CS:GO's mapchooser cooldown saving, because a map could pack a KV1 cfg with the string "ze_Mako". Then, when we go to save "ze_mako", it will instead be auto-converted to "ze_Mako" because that's what the server used first. So we are just re-implementing the checks to avoid this being an issue.