Lots of map system updates (#327)
* Add individual map cooldown configuration
List cooldown of all maps when running !nominate without having set a nomination

* Add map cooldown saving and loading from file

* Move processing of map cooldowns to after vote

This is the same behavour as the map vote system in CS:GO, where the current map will not be set on cooldown until after it is voted off

* Min and Max player requirements

* Update example maplist.cfg

* Add Minimum and Maximum player information to !nominate map list when requirements are not met

* Use snake case format for min/max player keyvalues in maplist.cfg

* Re-add !mapcooldowns chat command

* Change map cooldown file path

* Re-add cs2f_vote_maps_cooldown cvar as the default map cooldown

* Automatically remove nominations when min/max player count requirements not met

* Fix string format of console messages when loading map list

* Add shared function for player count

* Fix a possible edge case with KV1 capitalization handling

Basically, KV1 will cache whichever capitalization gets used first for a string, *globally*.

This previously caused issues with CS:GO's mapchooser cooldown saving, because a map could pack a KV1 cfg with the string "ze_Mako". Then, when we go to save "ze_mako", it will instead be auto-converted to "ze_Mako" because that's what the server used first. So we are just re-implementing the checks to avoid this being an issue.

* Fix nomination resetting not working

* Allow changing map by ID if admin already has console access

* Make !mapcooldowns print in order again

* Don't show disabled maps in !nominate map list

* Fix & update !setnextmap

* Allow immediate RTV's when enabled manually at map start

* Update GetNeededRTVCount to use GetOnlinePlayerCount

* Add more detailed error messages to !nominate

* Added warning on several map-related commands for non-unique queries

* Adjustments for logging

Improves behaviour with a logging backend.. still need to upstream that :)

---------

Co-authored-by: m-arcuri <124469923+m-arcuri@users.noreply.github.com>
Co-authored-by: Hichatu <realhichatu@gmail.com>