Stop removing internal cvar flags
These missing definitions were actually:

FCVAR_REFERENCE - Means cvar is a reference, usually used to get a cvar reference of a cvar registered in other module, and is temporary until the actual cvar was registered
FCVAR_INITIAL_SETVALUE - Is set for a first convar SetValue either with its default_value or with a value from a gameinfo. Mostly for callbacks to check for.
FCVAR_PERFORMING_CALLBACKS - Is set when cvar is calling to its callbacks from CallChangeCallback or CallGlobalChangeCallbacks, usually means if you set the value during these callbacks, its value would be queued and set when all callbacks were fired
FCVAR_COMMANDLINE_ENFORCED - If cvar was set via launch options it would not be reset on calls to ResetConVarsToDefaultValuesByFlag