Fix edge case where map end could be prematurely detected
No it's not a rounding error, I'm not sure what the issue actually is. But with the previous implementation, there seemed to be a 1 second period at the end of time left where the round could still restart, but we would detect the map being over. To account for this, the check now just has an extra second of leeway.